tests: refactor common functions to test_utils
